upgrade to nvim-cmp. nvim-compe is by the same author and is now deprecated.
first impression is that it is much easier to configure the way I want
with tab to cycle and enter to choose without adding a newline. Felt
way less hacky to do so.

I added the snippet plugin because I want to try configuring some
snippets. That is what I am planning to do next.

I haven't checked how this works with the lsp yet. That is a todo still.