remove buffer source from main cmp settings. I used to use this as a sort of minimal hacky (but fairly functional really) lsp for ruby where the actual lsps are not great but now I think it is getting in the way more than it is helping.
I also fixed the select true on cmp settings which means I always need
to hit tab to use cmp which makes it easier to ignore the
autocomplete. I hope I am not forgetting some other reason for this
setting but for now this seems better for my current uses.